+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/Hierarchy.java
@@ -559,33 +559,6 @@ public class Hierarchy {
     }
 
     /**
-     * Find XMethod for method in given list of classes, searching the classes
-     * in order.
-     *
-     * @param classList
-     *            list of classes in which to search
-     * @param methodName
-     *            the name of the method
-     * @param methodSig
-     *            the signature of the method
-     * @param chooser
-     *            JavaClassAndMethodChooser to select which methods are
-     *            considered; it must return true for a method to be returned
-     * @return the XMethod, or null if no such method exists in the class
-     */
-    @Deprecated
-    public static XMethod findXMethod(JavaClass[] classList, String methodName, String methodSig,
-            JavaClassAndMethodChooser chooser) {
-        for (JavaClass cls : classList) {
-            JavaClassAndMethod m;
-            if ((m = findMethod(cls, methodName, methodSig)) != null && chooser.choose(m)) {
-                return XFactory.createXMethod(cls, m.getMethod());
-            }
-        }
-        return null;
-    }
-
-    /**
      * Resolve possible method call targets. This works for both static and
      * instance method calls.
      *